pyproject.toml
[tool.poetry]
name = "softdesk"
version = "0.1.0"
description = "API RESTful for SoftDesk project management"
authors = ["devogs <billy.bob@devogs.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=5.2.1,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
djangorestframework-simplejwt = ">=5.5.0,<6.0.0"
drf-nested-routers = "^0.94.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
